+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Easy!Appointments - Changelog
 
+## [Unreleased]
+
+### Added
+
+- Update module support for Joomla 6
+
+
+
 ## [1.0.0] - 2024-10-29
 
 ### Added